Initial fixes for the polling transport within socketio.
This addresses a memory leak, and adds some monitoring by adding a thread to watch the sizes of the core maps.
ResponseLogic provides a "unified" error handler that adds the appropriate CORS headers.
PollingTransport should send an error if the first request is a GET.
AuthorizeHandler now will send an UNKNOWN_SID if one is provided but not actually in cache.
The downside to this is this check has to be made at every layer in the pipeline, so we will still see "error" log messages in the case of a race condition.
Added a simple "ClientHead" channel retrieval method.